sync: Add waitgroup module implementation inspired by Go language
This commit adds the implementation of waitgroup module, which is inspired by Go language.
The waitgroup module provides three methods: add, done, and wait.

It allows the synchronization of coroutines by waiting for them to complete their tasks.
The waitgroup module is implemented using a condition variable, and it can be used to
wait for a specified number of coroutines to complete their tasks.

Signed-off-by: Jianhui Zhao <zhaojh329@gmail.com>